Added Collision
Changed GameObject variables to double
Added Debug Class
Fixed collision functions (I hope)

+++ src/com/game/framework/Camera.java
@@ -9,8 +9,8 @@ public class Camera {
 	public void tick(){
 		
 		if(pl != null){
-			setX(-pl.getX() + Game.getScreenWidth()/2);	
-			setY(-pl.getY() + Game.getScreenHeight()/2);
+			setX((int)-pl.getX() + Game.getScreenWidth()/2);	
+			setY((int)-pl.getY() + Game.getScreenHeight()/2);
 		}
 	}
